code/api-requirements.txt
aiofiles==0.6.0
alchemy-mock==0.4.3
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
coverage==5.5
docutils==0.17.1
fastapi==0.61.2
feedparser==6.0.2
future==0.18.2
fuzzysearch==0.7.3
fuzzywuzzy==0.18.0
h11==0.11.0
idna==2.10
iniconfig==1.1.1
logdna==1.5.3
lxml==4.6.2
numpy==1.19.4
oauthlib==3.1.0
packaging==20.9
pandas==1.1.4
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pydantic==1.7.2
pyparsing==2.4.7
PyMuPDF==1.18.16
pytest==6.2.4
pytest-cov==2.12.1
pytest-dotenv==0.5.2
pytest-mock==3.6.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-Levenshtein==0.12.2
python-multipart==0.0.5
python-twitter==3.5
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.12.3
watson-developer-cloud==2.8.1
websocket-client==0.48.0
xlrd==1.2.0
XlsxWriter==1.3.7
tenacity==8.0.1
